Fix types in tests && change test filenames
