+++ tools/data-handler/test/command-fetch.test.ts
@@ -6,7 +6,7 @@ import { mkdirSync, rmSync } from 'node:fs';
 import { join } from 'node:path';
 
 // cyberismo
-import { type CardsOptions, Cmd, Commands } from '../src/command-handler.js';
+import { Cmd, Commands } from '../src/command-handler.js';
 import { copyDir } from '../src/utils/file-utils.js';
 
 // Create test artifacts in a temp folder.
@@ -16,7 +16,7 @@ const testDir = join(baseDir, 'tmp-command-handler-fetch-tests');
 const decisionRecordsPath = join(testDir, 'valid/decision-records');
 
 const commandHandler: Commands = new Commands();
-const options: CardsOptions = { projectPath: decisionRecordsPath };
+const options = { projectPath: decisionRecordsPath };
 
 describe('fetch command', () => {
   before(async () => {
