+++ tools/data-handler/test/command-rank.test.ts
@@ -7,7 +7,7 @@ import { join } from 'node:path';
 
 // cyberismo
 import { copyDir } from '../src/utils/file-utils.js';
-import { type CardsOptions, Cmd, Commands } from '../src/command-handler.js';
+import { Cmd, Commands } from '../src/command-handler.js';
 import { Project } from '../src/containers/project.js';
 import { Show } from '../src/commands/index.js';
 
@@ -18,7 +18,7 @@ const testDir = join(baseDir, 'tmp-command-handler-rank-tests');
 const decisionRecordsPath = join(testDir, 'valid/decision-records');
 
 const commandHandler: Commands = new Commands();
-const options: CardsOptions = { projectPath: decisionRecordsPath };
+const options = { projectPath: decisionRecordsPath };
 
 describe('rank command', () => {
   let rootCardKey: string;
