+++ tools/data-handler/test/command-validate.test.ts
